Add integration test for HTTPClient
A new project for the integration tests were created so we can the
unit tests without having to setup the dependencies for the integration
tests. The HTTPClient test needs wiremock running before start or
otherwise will fail.
Also was added a new exception for duplicated migration, if that
exception happens the migration is stored as ran, in the future an
option for that will be created