Implement integration test for the database
The tests are a bit clunky because the framework don't have a method
like BeforeAll in JUnit, so for now we are connecting and disconnecting
to the database every test, I will look for a solution later.

The only method that is not tested yet is the Commit, I will come back
later for that.